# public/openapi.yaml
info:
  description: |
    Purpie is an open source social media solution focusing on video oriented communication and content sharing.
    
    The Purpie endpoints are offered to our users to facilitate the use of our technology and allow seamless integrations with their applications.
    For this documentation, you can use the **Api Key** and **Api Secret** which generated from [Purpie](https://purpie.io/) settings page to test the services.

    # Introduction
    Purpie orchestrates various media services and solutions for online meetings, live streamings, on demand video delivery and many more.
    With an agile and hierarchical way of user interaction and content sharing possibilities, Purpie proposes a robust and flexible solution for
    various video communication and distribution needs.

    # Cross-Origin Resource Sharing
    This API features Cross-Origin Resource Sharing (CORS) implemented.
    And that allows cross-domain communication from the browser.
    All responses have a wildcard same-origin which makes them completely public and accessible to everyone, including any code on any site.

    # Authentication
    To make authorized calls on Purpie API operations, you need to generate an API key and secret in the settings page after login to [Purpie](https://purpie.io/).
    
    After generating API credentials, you need to note your `api secret`. You can access this field once after creating. If you lost or forgot your api
...
    
...
    header:
    
...

...
  x-logo:
...
servers:
...
openapi: 3.0.0
paths:
...
  /v1/user/contact/invitation/response:
    post:
      operationId: UserController_contactInvitationResponse
      summary: Respond Contact Invitation
      description: Respond contact invitation with request payload.
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContactInvitationResponseDto'
      responses:
        '201':
          description: User responds to contact invitation
          content:
            application/json:
              schema:
                type: string
                example: OK
        '400':
          description: Error thrown when the requested contact id is same as authenticated user id.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 404
                  message:
                    type: string
                    example: You cannot add yourself to your contacts
                  error:
                    type: string
                    example: CONTACT_ELIGIBILITY_ERROR
        '401':
          description: Error thrown when the user has not authorized or JWT is invalid.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: You not authorized to use this route
                  error:
                    type: string
                    example: NOT_SIGNED_IN
        '404':
          description: Error thrown when the requested contact invitation not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 404
                  message:
                    type: string
                    example: Contact Invitation not found
                  error:
                    type: string
                    example: CONTACT_INVITATION_NOT_FOUND
      tags:
        - User
  /v1/user/contact/invitation/create:
    post:
      operationId: UserController_createNewContactInvitation
      summary: Create Contact Invitation
      description: Authenticated user creates a new contact invitation for a user belonging to email.
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContactDto'
      responses:
        '201':
          description: User creates a new contact invitation
          content:
            application/json:
              schema:
                type: integer
        '400':
          description: Error thrown when the user already sent invitation to requested email.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 400
                  message:
                    type: string
                    example: Invitation to this user has already been sent
                  error:
                    type: string
                    example: INVITATION_ALREADY_SENT_FOR_USER
        '401':
          description: Error thrown when the user has not authorized or JWT is invalid.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: You not authorized to use this route
                  error:
                    type: string
                    example: NOT_SIGNED_IN
        '409':
          description: Error thrown when the user already received invitation .
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 400
                  message:
                    type: string
                    example: You have already been invited by this user
                  error:
                    type: string
                    example: INVITATION_ALREADY_RECEIVED_FROM_USER
      tags:
        - User
  /v1/user/invitations/list:
    get:
      operationId: UserController_getUserInvitations
      summary: List Invitations
      description: User lists all invitations, contacts, channels, zones.
      parameters:
        - name: limit
          required: false
          in: query
          description: The number of records to get.
          schema:
            default: 30
            type: number
        - name: skip
          required: false
          in: query
          description: The number of records to skip.
          schema:
            default: 0
            type: number
      responses:
        '200':
          description: Lists all invitations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactInvitationListResponse'
        '401':
          description: Error thrown when the user has not authorized or JWT is invalid.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: You not authorized to use this route
                  error:
                    type: string
                    example: NOT_SIGNED_IN
      tags:
        - User
  /v1/user/contact/list:
    get:
      operationId: UserController_listContacts
      summary: List Contacts
      description: User lists contacts.
      parameters:
        - name: limit
          required: false
          in: query
          description: The number of records to get.
          schema:
            default: 30
            type: number
        - name: skip
          required: false
          in: query
          description: The number of records to skip.
          schema:
            default: 0
            type: number
      responses:
        '200':
          description: User lists contacts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactListResponse'
        '401':
          description: Error thrown when the user has not authorized or JWT is invalid.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: You not authorized to use this route
                  error:
                    type: string
                    example: NOT_SIGNED_IN
      tags:
        - User
  /v1/user/contact/list/{userName}:
    get:
      operationId: UserController_listPublicUserContacts
      summary: List User Contact
      description: User lists another public users contacts
      parameters:
        - name: userName
          required: true
          in: path
          schema:
            type: string
        - name: limit
          required: false
          in: query
          description: The number of records to get.
          schema:
            default: 30
            type: number
        - name: skip
          required: false
          in: query
          description: The number of records to skip.
          schema:
            default: 0
            type: number
      responses:
        '200':
          description: User lists another public users contacts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactListResponse'
        '401':
          description: Error thrown when the user has not authorized or JWT is invalid.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: You not authorized to use this route
                  error:
                    type: string
                    example: NOT_SIGNED_IN
      tags:
        - User
  /v1/user/contact/remove/{contactId}:
    delete:
      operationId: UserController_deleteContact
      summary: Delete Contact
      description: User deletes a contact with requested "contactId."
      parameters:
        - name: contactId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: User removes a contact by id
          content:
            application/json:
              schema:
                type: string
                example: OK
        '401':
          description: Error thrown when the user has not authorized or JWT is invalid.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: You not authorized to use this route
                  error:
                    type: string
                    example: NOT_SIGNED_IN
      tags:
        - User
  /v1/user/search:
    get:
      operationId: UserController_searchUsers
      summary: Search Users
      description: Search user with requested payload.
      parameters:
        - name: limit
          required: false
          in: query
          description: The number of records to get.
          schema:
            default: 30
            type: number
        - name: skip
          required: false
          in: query
          description: The number of records to skip.
          schema:
            default: 0
            type: number
        - name: name
          required: true
          in: query
          schema:
            type: string
        - name: excludeIds
          required: false
          in: query
          description: User ids to exclude while searching. Should be a comma separated ids
          schema:
            type: string
        - name: channelId
          required: false
          in: query
          schema:
            type: string
        - name: userContacts
          required: false
          in: query
          description: specify true to search in current user's contacts
          schema:
            type: string
        - name: excludeCurrentUser
          required: false
          in: query
          description: Exclude current user in search. Specify false to disable this.
          schema:
            type: string
      responses:
        '200':
          description: Search user with requested payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSearchResponse'
        '401':
          description: Error thrown when the user has not authorized or JWT is invalid.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: You not authorized to use this route
                  error:
                    type: string
                    example: NOT_SIGNED_IN
      tags:
        - User
  /v1/user/role/change:
    put:
      operationId: UserController_changeUserRole
      summary: Update User Role
      description: User changes a role for an existing user. User must have "canManageRole" permission
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetUserRoleDto'
      responses:
        '201':
          description: Update user role
          content:
            application/json:
              schema:
                type: string
                example: OK
        '400':
          description: Error thrown when validation error occurs according to requested payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Error thrown when the user has not authorized or JWT is invalid.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: You not authorized to use this route
                  error:
                    type: string
                    example: NOT_SIGNED_IN
        '403':
          description: Error thrown when changed user role is last super admin.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 403
                  message:
                    type: string
                    example: There must be at least one super admin
                  error:
                    type: string
                    example: SUPER_ADMIN_NOT_EXIST
      tags:
        - User
  /v1/user/user-name-check:
    post:
      operationId: UserController_userNameExistenceCheck
      summary: Check Username
      description: Checks if user name provided has already been taken.
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserNameExistenceCheckDto'
      responses:
        '201':
          description: Checked username
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserNameExistenceCheckResponse'
        '400':
          description: Error thrown when validation error occurs according to requested payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
      tags:
        - User
  /v1/user/zone/list/{userName}:
    get:
      operationId: UserController_listPublicUserZones
      summary: List User Zones
      description: List public zones of user belonging to "userName" parameter.
      parameters:
        - name: userName
          required: true
          in: path
          schema:
            type: string
        - name: limit
          required: false
          in: query
          description: The number of records to get.
          schema:
            default: 30
            type: number
        - name: skip
          required: false
          in: query
          description: The number of records to skip.
          schema:
            default: 0
            type: number
      responses:
        '200':
          description: List requested user public zones
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserZoneResponse'
        '401':
          description: Error thrown when the user has not authorized or JWT is invalid.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: You not authorized to use this route
                  error:
                    type: string
                    example: NOT_SIGNED_IN
      tags:
        - User
...
  /v1/user/display-photo/{fileName}:
    get:
      operationId: UserController_viewProfilePhoto
      summary: Get Display Photo
      description: View display photo of authenticated user.
      parameters:
        - name: fileName
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: View display photo of authenticated user
      tags:
        - User
...
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    ContactInvitationResponseDto:
      type: object
      properties:
        status:
          type: string
          enum:
            - accept
            - reject
        contactInvitationId:
          type: string
      required:
        - status
        - contactInvitationId
    CreateContactDto:
      type: object
      properties:
        email:
          type: string
      required:
        - email
...